Muunnettu hakijaQuestionGroupSpec mocha-testi playwright-testiksi
