--- resources/templates/hakija-question-group-form-test.html
@@ -1,24 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head lang="en">
-  <meta charset="UTF-8"/>
-  <title>UI tests</title>
-</head>
-<body>
-<iframe id="test"></iframe>
-<div id="mocha" class="foo"></div>
-<script src="//cdnjs.cloudflare.com/ajax/libs/mocha/2.5.3/mocha.min.js"></script>
-<script src="//cdnjs.cloudflare.com/ajax/libs/chai/3.5.0/chai.js"></script>
-<script src="//cdnjs.cloudflare.com/ajax/libs/lodash.js/4.13.1/lodash.min.js"></script>
-<script src="//cdnjs.cloudflare.com/ajax/libs/q.js/1.4.1/q.min.js"></script>
-<script src="//code.jquery.com/jquery-2.2.4.min.js"></script>
-
-<link rel="stylesheet" href="/hakemus/css/compiled/mocha.css">
-
-<script src="/hakemus/spec/util.js"></script>
-<script src="/hakemus/spec/hakijaCommon.js"></script>
-<script src="/hakemus/spec/hakijaQuestionGroupSpec.js"></script>
-<script src="/hakemus/spec/testRunner.js"> </script>
-
-</body>
-</html>
